Changes summed below:
1. Mutexes for writing and publishing new states
2. Changed data type of increment counters to check pub interval of depth and video
3. Times stamps are synced for all objects based on the simulation physics update

+++ ambf_plugins/core/ros_comm_plugin/VideoStreamerPlugin.h
@@ -67,11 +67,11 @@ private:
 
     // Counter for the times we have written to ambf_comm API
     // This is only for internal use as it could be reset
-    unsigned short m_write_count = 0;
+    unsigned int m_write_count = 0;
 
     // Counter for the times we have read from ambf_comm API
     // This is only for internal use as it could be reset
-    unsigned short m_read_count = 0;
+    unsigned int m_read_count = 0;
     // Image Transport ROS Node
     ros::NodeHandle* m_rosNode;
     afCameraPtr m_cameraPtr = nullptr;
